+++ tests/fixtures/cassettes/test_a_persistent_invalid_xml_error_is_raised.yml
@@ -6,9 +6,10 @@ interactions:
       - flickr-photos-api/dev (https://github.com/Flickr-Foundation/flickr-photos-api;
         hello@flickr.org)
     method: GET
-    uri: https://api.flickr.com/services/rest/?method=flickr.urls.lookupUser&url=https%3A%2F%2Fwww.flickr.com%2Fphotos%2Fnavymedicine%2F
+    uri: https://api.flickr.com/services/rest/?extras=license%2Cdate_upload%2Cdate_taken%2Cmedia%2Coriginal_format%2Cowner_name%2Curl_sq%2Curl_t%2Curl_s%2Curl_m%2Curl_o%2Ctags%2Cgeo%2Curl_q%2Curl_l%2Cdescription%2Csafety_level%2Crealname%2Cpath_alias%2Ccount_comments%2Ccount_views&method=flickr.people.getPublicPhotos&page=1&per_page=10&user_id=61270229%40N05
   response:
-    content: "{\"stat\":\"fail\",\"message\":\"Sorry, the Flickr API service is not currently available.\",\"code\":201},"
+    content: '{"stat":"fail","message":"Sorry, the Flickr API service is not currently
+      available.","code":201},'
     headers:
       Content-Type:
       - text/xml; charset=utf-8
@@ -21,9 +22,10 @@ interactions:
       - flickr-photos-api/dev (https://github.com/Flickr-Foundation/flickr-photos-api;
         hello@flickr.org)
     method: GET
-    uri: https://api.flickr.com/services/rest/?method=flickr.urls.lookupUser&url=https%3A%2F%2Fwww.flickr.com%2Fphotos%2Fnavymedicine%2F
+    uri: https://api.flickr.com/services/rest/?extras=license%2Cdate_upload%2Cdate_taken%2Cmedia%2Coriginal_format%2Cowner_name%2Curl_sq%2Curl_t%2Curl_s%2Curl_m%2Curl_o%2Ctags%2Cgeo%2Curl_q%2Curl_l%2Cdescription%2Csafety_level%2Crealname%2Cpath_alias%2Ccount_comments%2Ccount_views&method=flickr.people.getPublicPhotos&page=1&per_page=10&user_id=61270229%40N05
   response:
-    content: "{\"stat\":\"fail\",\"message\":\"Sorry, the Flickr API service is not currently available.\",\"code\":201},"
+    content: '{"stat":"fail","message":"Sorry, the Flickr API service is not currently
+      available.","code":201},'
     headers:
       Content-Type:
       - text/xml; charset=utf-8
@@ -36,9 +38,10 @@ interactions:
       - flickr-photos-api/dev (https://github.com/Flickr-Foundation/flickr-photos-api;
         hello@flickr.org)
     method: GET
-    uri: https://api.flickr.com/services/rest/?method=flickr.urls.lookupUser&url=https%3A%2F%2Fwww.flickr.com%2Fphotos%2Fnavymedicine%2F
+    uri: https://api.flickr.com/services/rest/?extras=license%2Cdate_upload%2Cdate_taken%2Cmedia%2Coriginal_format%2Cowner_name%2Curl_sq%2Curl_t%2Curl_s%2Curl_m%2Curl_o%2Ctags%2Cgeo%2Curl_q%2Curl_l%2Cdescription%2Csafety_level%2Crealname%2Cpath_alias%2Ccount_comments%2Ccount_views&method=flickr.people.getPublicPhotos&page=1&per_page=10&user_id=61270229%40N05
   response:
-    content: "{\"stat\":\"fail\",\"message\":\"Sorry, the Flickr API service is not currently available.\",\"code\":201},"
+    content: '{"stat":"fail","message":"Sorry, the Flickr API service is not currently
+      available.","code":201},'
     headers:
       Content-Type:
       - text/xml; charset=utf-8
@@ -51,9 +54,10 @@ interactions:
       - flickr-photos-api/dev (https://github.com/Flickr-Foundation/flickr-photos-api;
         hello@flickr.org)
     method: GET
-    uri: https://api.flickr.com/services/rest/?method=flickr.urls.lookupUser&url=https%3A%2F%2Fwww.flickr.com%2Fphotos%2Fnavymedicine%2F
+    uri: https://api.flickr.com/services/rest/?extras=license%2Cdate_upload%2Cdate_taken%2Cmedia%2Coriginal_format%2Cowner_name%2Curl_sq%2Curl_t%2Curl_s%2Curl_m%2Curl_o%2Ctags%2Cgeo%2Curl_q%2Curl_l%2Cdescription%2Csafety_level%2Crealname%2Cpath_alias%2Ccount_comments%2Ccount_views&method=flickr.people.getPublicPhotos&page=1&per_page=10&user_id=61270229%40N05
   response:
-    content: "{\"stat\":\"fail\",\"message\":\"Sorry, the Flickr API service is not currently available.\",\"code\":201},"
+    content: '{"stat":"fail","message":"Sorry, the Flickr API service is not currently
+      available.","code":201},'
     headers:
       Content-Type:
       - text/xml; charset=utf-8
@@ -66,9 +70,10 @@ interactions:
       - flickr-photos-api/dev (https://github.com/Flickr-Foundation/flickr-photos-api;
         hello@flickr.org)
     method: GET
-    uri: https://api.flickr.com/services/rest/?method=flickr.urls.lookupUser&url=https%3A%2F%2Fwww.flickr.com%2Fphotos%2Fnavymedicine%2F
+    uri: https://api.flickr.com/services/rest/?extras=license%2Cdate_upload%2Cdate_taken%2Cmedia%2Coriginal_format%2Cowner_name%2Curl_sq%2Curl_t%2Curl_s%2Curl_m%2Curl_o%2Ctags%2Cgeo%2Curl_q%2Curl_l%2Cdescription%2Csafety_level%2Crealname%2Cpath_alias%2Ccount_comments%2Ccount_views&method=flickr.people.getPublicPhotos&page=1&per_page=10&user_id=61270229%40N05
   response:
-    content: "{\"stat\":\"fail\",\"message\":\"Sorry, the Flickr API service is not currently available.\",\"code\":201},"
+    content: '{"stat":"fail","message":"Sorry, the Flickr API service is not currently
+      available.","code":201},'
     headers:
       Content-Type:
       - text/xml; charset=utf-8
@@ -81,9 +86,10 @@ interactions:
       - flickr-photos-api/dev (https://github.com/Flickr-Foundation/flickr-photos-api;
         hello@flickr.org)
     method: GET
-    uri: https://api.flickr.com/services/rest/?method=flickr.urls.lookupUser&url=https%3A%2F%2Fwww.flickr.com%2Fphotos%2Fnavymedicine%2F
+    uri: https://api.flickr.com/services/rest/?extras=license%2Cdate_upload%2Cdate_taken%2Cmedia%2Coriginal_format%2Cowner_name%2Curl_sq%2Curl_t%2Curl_s%2Curl_m%2Curl_o%2Ctags%2Cgeo%2Curl_q%2Curl_l%2Cdescription%2Csafety_level%2Crealname%2Cpath_alias%2Ccount_comments%2Ccount_views&method=flickr.people.getPublicPhotos&page=1&per_page=10&user_id=61270229%40N05
   response:
-    content: "{\"stat\":\"fail\",\"message\":\"Sorry, the Flickr API service is not currently available.\",\"code\":201},"
+    content: '{"stat":"fail","message":"Sorry, the Flickr API service is not currently
+      available.","code":201},'
     headers:
       Content-Type:
       - text/xml; charset=utf-8
