+++ tests/fixtures/cassettes/test_get_photos_from_flickr_url[single_photo].yml
@@ -65,13 +65,13 @@ interactions:
       Content-Type:
       - text/xml; charset=utf-8
       Date:
-      - Mon, 29 Apr 2024 14:05:01 GMT
+      - Mon, 29 Apr 2024 16:49:52 GMT
       Transfer-Encoding:
       - chunked
       Via:
-      - 1.1 5a98aec7da8cddc4fee7bc85f8beb31a.cloudfront.net (CloudFront)
+      - 1.1 5e8927f6dbbe16e857124daf8548aeb2.cloudfront.net (CloudFront)
       X-Amz-Cf-Id:
-      - p0LRSPHFa6z0W1TehRlAjbyDD5nK1DXMWwOJujQdeieGEK_clkSsyw==
+      - vBBBFZhxbbVwtOViM7r9Qy3qFJUxgXjLFBrCkBDNLT9bisBPIWTZ2w==
       X-Amz-Cf-Pop:
       - LHR5-P1
       X-Cache:
@@ -82,17 +82,17 @@ interactions:
       - openresty
       set-cookie:
       - ccc=%7B%22needsConsent%22%3Atrue%2C%22managed%22%3A0%2C%22changed%22%3A0%2C%22info%22%3A%7B%22cookieBlock%22%3A%7B%22level%22%3A0%2C%22blockRan%22%3A0%7D%7D%7D;
-        expires=Wed, 29-May-2024 14:05:01 GMT; Max-Age=2592000; path=/; domain=.flickr.com
+        expires=Wed, 29-May-2024 16:49:52 GMT; Max-Age=2592000; path=/; domain=.flickr.com
       - ccc=%7B%22needsConsent%22%3Atrue%2C%22managed%22%3A0%2C%22changed%22%3A0%2C%22info%22%3A%7B%22cookieBlock%22%3A%7B%22level%22%3A0%2C%22blockRan%22%3A1%7D%7D%7D;
-        expires=Wed, 29-May-2024 14:05:01 GMT; Max-Age=2592000; path=/; domain=.flickr.com
+        expires=Wed, 29-May-2024 16:49:52 GMT; Max-Age=2592000; path=/; domain=.flickr.com
       vary:
       - Accept-Encoding
       x-flickr-api-request:
-      - Self=1-662fa90d-5dc626796cab07573de197f9;Root=1-662fa90d-0ab22e1f67f47daf700ed5ce
+      - Self=1-662fcfb0-002136fc4ff74f193433c54b;Root=1-662fcfb0-4f5f69a9185a2c99178dcf6b
       x-robots-tag:
       - noindex
       x-server:
-      - serverless-proxy-10.78.26.197
+      - serverless-proxy-10.78.16.171
     http_version: HTTP/1.1
     status_code: 200
 - request:
@@ -135,13 +135,13 @@ interactions:
       Content-Type:
       - text/xml; charset=utf-8
       Date:
-      - Mon, 29 Apr 2024 14:05:01 GMT
+      - Mon, 29 Apr 2024 16:49:52 GMT
       Transfer-Encoding:
       - chunked
       Via:
-      - 1.1 5a98aec7da8cddc4fee7bc85f8beb31a.cloudfront.net (CloudFront)
+      - 1.1 5e8927f6dbbe16e857124daf8548aeb2.cloudfront.net (CloudFront)
       X-Amz-Cf-Id:
-      - WoDufE7axS9ojLocFEuAeap9FJ405_fHNrxW_b98yOF1mGJ-FMxICA==
+      - rBPx-45MIVxi3fp6PrnCAZptmat3DbCEpjJulEMb_eclF9FWyEIrnQ==
       X-Amz-Cf-Pop:
       - LHR5-P1
       X-Cache:
@@ -152,15 +152,15 @@ interactions:
       - openresty
       set-cookie:
       - ccc=%7B%22needsConsent%22%3Atrue%2C%22managed%22%3A0%2C%22changed%22%3A0%2C%22info%22%3A%7B%22cookieBlock%22%3A%7B%22level%22%3A0%2C%22blockRan%22%3A1%7D%7D%7D;
-        expires=Wed, 29-May-2024 14:05:01 GMT; Max-Age=2592000; path=/; domain=.flickr.com
+        expires=Wed, 29-May-2024 16:49:52 GMT; Max-Age=2592000; path=/; domain=.flickr.com
       vary:
       - Accept-Encoding
       x-flickr-api-request:
-      - Self=1-662fa90d-2e61b7ac58a1562954e061c5;Root=1-662fa90d-36ebba850458f34d07a79620
+      - Self=1-662fcfb0-2f8f756a70f41f077ed1820b;Root=1-662fcfb0-71265fcd175dc37f2ddae5ce
       x-robots-tag:
       - noindex
       x-server:
-      - serverless-proxy-10.78.11.164
+      - serverless-proxy-10.78.16.171
     http_version: HTTP/1.1
     status_code: 200
 - request:
@@ -217,13 +217,13 @@ interactions:
       Content-Type:
       - text/xml; charset=utf-8
       Date:
-      - Mon, 29 Apr 2024 14:05:01 GMT
+      - Mon, 29 Apr 2024 16:49:52 GMT
       Transfer-Encoding:
       - chunked
       Via:
-      - 1.1 5a98aec7da8cddc4fee7bc85f8beb31a.cloudfront.net (CloudFront)
+      - 1.1 5e8927f6dbbe16e857124daf8548aeb2.cloudfront.net (CloudFront)
       X-Amz-Cf-Id:
-      - Xpy2u8hlcK04D0u-uCcvdSvcPVfRlMGDgDlG1fN3-WBI1H3jzGLHsw==
+      - rysUOx7qFLay7Mde9AFILBRLDxpl3-dqsjvLjC18pfPBSBbimrGs7g==
       X-Amz-Cf-Pop:
       - LHR5-P1
       X-Cache:
@@ -234,11 +234,11 @@ interactions:
       - openresty
       set-cookie:
       - ccc=%7B%22needsConsent%22%3Atrue%2C%22managed%22%3A0%2C%22changed%22%3A0%2C%22info%22%3A%7B%22cookieBlock%22%3A%7B%22level%22%3A0%2C%22blockRan%22%3A1%7D%7D%7D;
-        expires=Wed, 29-May-2024 14:05:01 GMT; Max-Age=2592000; path=/; domain=.flickr.com
+        expires=Wed, 29-May-2024 16:49:52 GMT; Max-Age=2592000; path=/; domain=.flickr.com
       vary:
       - Accept-Encoding
       x-flickr-api-request:
-      - Self=1-662fa90d-2e51aa0c3f6bcfb94abc2bdd;Root=1-662fa90d-187ce2b9642c17ab7acfe493
+      - Self=1-662fcfb0-017d36362e82447a64f7065e;Root=1-662fcfb0-3c7ca817218cc7936ebf7693
       x-robots-tag:
       - noindex
       x-server:
