+++ tests/fixtures/cassettes/test_methods_fail_if_not_found[get_photos_in_album_with_missing_user].yml
@@ -14,23 +14,23 @@ interactions:
       - flickr-photos-api/dev (https://github.com/Flickr-Foundation/flickr-photos-api;
         hello@flickr.org)
     method: GET
-    uri: https://api.flickr.com/services/rest/?method=flickr.urls.lookupUser&url=https%3A%2F%2Fwww.flickr.com%2Fphotos%2FDefinitelyDoesNotExist
+    uri: https://api.flickr.com/services/rest/?extras=license%2Cdate_upload%2Cdate_taken%2Cmedia%2Coriginal_format%2Cowner_name%2Curl_sq%2Curl_t%2Curl_s%2Curl_m%2Curl_o%2Ctags%2Cgeo%2Curl_q%2Curl_l%2Cdescription%2Csafety_level%2Crealname%2Cpath_alias%2Ccount_comments%2Ccount_views&method=flickr.photosets.getPhotos&page=1&per_page=10&photoset_id=1234&user_id=-1
   response:
     content: "<?xml version=\"1.0\" encoding=\"utf-8\" ?>\n<rsp stat=\"fail\">\n\t<err
-      code=\"1\" msg=\"User not found\" />\n</rsp>\n"
+      code=\"2\" msg=\"User not found\" />\n</rsp>\n"
     headers:
       Connection:
       - keep-alive
       Content-Type:
       - text/xml; charset=utf-8
       Date:
-      - Mon, 29 Apr 2024 14:05:45 GMT
+      - Mon, 29 Apr 2024 17:25:04 GMT
       Transfer-Encoding:
       - chunked
       Via:
-      - 1.1 f9b4eb435f0b621adc8e78b8d2ac6e70.cloudfront.net (CloudFront)
+      - 1.1 cebcb0d17c62ea30e361587bfe114ca0.cloudfront.net (CloudFront)
       X-Amz-Cf-Id:
-      - dCf6TIfA-uUJfVREbN47Ow8vMCnNl61OJtIC74ZoYJr1GXnlW9IbRg==
+      - q50BpuwCXivcZrb3mCKQqcawwoOOefhAGfNxi1apms0h7v7Tehov-g==
       X-Amz-Cf-Pop:
       - LHR5-P1
       X-Cache:
@@ -41,17 +41,17 @@ interactions:
       - openresty
       set-cookie:
       - ccc=%7B%22needsConsent%22%3Atrue%2C%22managed%22%3A0%2C%22changed%22%3A0%2C%22info%22%3A%7B%22cookieBlock%22%3A%7B%22level%22%3A0%2C%22blockRan%22%3A0%7D%7D%7D;
-        expires=Wed, 29-May-2024 14:05:44 GMT; Max-Age=2592000; path=/; domain=.flickr.com
+        expires=Wed, 29-May-2024 17:25:04 GMT; Max-Age=2592000; path=/; domain=.flickr.com
       - ccc=%7B%22needsConsent%22%3Atrue%2C%22managed%22%3A0%2C%22changed%22%3A0%2C%22info%22%3A%7B%22cookieBlock%22%3A%7B%22level%22%3A0%2C%22blockRan%22%3A1%7D%7D%7D;
-        expires=Wed, 29-May-2024 14:05:44 GMT; Max-Age=2592000; path=/; domain=.flickr.com
+        expires=Wed, 29-May-2024 17:25:04 GMT; Max-Age=2592000; path=/; domain=.flickr.com
       vary:
       - Accept-Encoding
       x-flickr-api-request:
-      - Self=1-662fa938-5b43b8ef603434bf408a37e0;Root=1-662fa938-4e6266c5317fac5123de0251
+      - Self=1-662fd7f0-39d501853a81291207c79a56;Root=1-662fd7f0-0df864bb506bb64a60ebc94c
       x-robots-tag:
       - noindex
       x-server:
-      - serverless-proxy-10.78.44.190
+      - serverless-proxy-10.78.13.106
     http_version: HTTP/1.1
     status_code: 200
 version: 1
